example shows bug with let's (which should be one word)
