+++ Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "words-rs"
+name = "words"
 version = "0.1.0"
 edition = "2021"
 author = "Sarah Allen"
