+++ src/lib.rs
@@ -1,0 +1,15 @@
+use regex::Regex;
+
+pub fn html_words<S: AsRef<str>>(text: S) -> anyhow::Result<String> {
+    let regex = Regex::new(r"([a-zà-ýA-ZÀ-Ý0-9]+?)([[\s$][^a-zà-ýA-ZÀ-Ý0-9]]+)")?;
+    let mut nth_word = 0;
+    let html_string = regex.captures_iter(text.as_ref()).map(|c| {
+        println!("{:?}", c);
+        let range: std::ops::Range<usize> = c.get(0).unwrap().range();
+        let s = format!("<span word='{}' char='{}'>{}</span>{}", 
+                        nth_word, range.start, &c[1], &c[2]);
+        nth_word = nth_word + 1;
+        s
+    }).collect::<Vec<String>>().join("");
+    Ok(html_string)    
+}
